==> C. Gather travel behaviour data/17. Part of day LMS.ipynb <==
import pandas as pd
import geopandas as gpd
import numpy as np
import matplotlib.pyplot as plt

import os


%matplotlib inline

parent_dir = os.path.split(os.getcwd())[0]

zones = gpd.read_file(parent_dir + '\\Data\\New\\lms_zone_du_new.shp') # LMS Zone data

ovin = pd.read_csv(parent_dir + '\\Data\\New\\Ovin_final.csv', index_col=0)

lms_orig = pd.read_csv(parent_dir + '\\Data\\New\\lms_modal_split_orig_abs.csv', index_col=0)

no_list = ['21', '31', '51']

car_list = []

for s in no_list:
    path = '\\Data\\LMS\\HB_matrices\\AUTO\\TRP2' + s + '18_synth.txt'
    car_list.append(pd.read_csv(parent_dir + path, delimiter='\t'))


car_final = car_list[0].copy()

for i in range(1, len(car_list)):
    car_list[i] = car_list[i].rename(columns={'VAR':i})
    car_final = car_final.merge(car_list[i], on=['Orig', 'Dest'], how='outer')

car_final['tot'] = car_final.iloc[:, 2:].sum(axis=1)
car_final = car_final[['Orig', 'Dest', 'tot']]

no_list = ['23', '33', '53']

car_list = []

for s in no_list:
    path = '\\Data\\LMS\\HB_matrices\\AUTO\\TRP2' + s + '18_synth.txt'
    car_list.append(pd.read_csv(parent_dir + path, delimiter='\t'))


car_final3 = car_list[0].copy()

for i in range(1, len(car_list)):
    car_list[i] = car_list[i].rename(columns={'VAR':i})
    car_final3 = car_final3.merge(car_list[i], on=['Orig', 'Dest'], how='outer')

car_final3['tot'] = car_final3.iloc[:, 2:].sum(axis=1)
car_final3 = car_final3[['Orig', 'Dest', 'tot']]

no_list = ['22', '32', '52']

car_list = []

for s in no_list:
    path = '\\Data\\LMS\\HB_matrices\\AUTO\\TRP2' + s + '18_synth.txt'
    car_list.append(pd.read_csv(parent_dir + path, delimiter='\t'))


car_final2 = car_list[0].copy()

for i in range(1, len(car_list)):
    car_list[i] = car_list[i].rename(columns={'VAR':i})
    car_final2 = car_final2.merge(car_list[i], on=['Orig', 'Dest'], how='outer')

car_final2['tot'] = car_final2.iloc[:, 2:].sum(axis=1)
car_final2 = car_final2[['Orig', 'Dest', 'tot']]

car_driver = pd.read_csv(parent_dir + '\\Data\\New\\LMS_results\\lms-car-driver', index_col=0)

btm1 = pd.read_csv(parent_dir + '\\Data\\LMS\\HB_matrices\\BTM\\TRP9_118.dat', delimiter='\t')
btm2 = pd.read_csv(parent_dir + '\\Data\\LMS\\HB_matrices\\BTM\\TRP9_218.dat', delimiter='\t')
btm3 = pd.read_csv(parent_dir + '\\Data\\LMS\\HB_matrices\\BTM\\TRP9_318.dat', delimiter='\t')


trein1 = pd.read_csv(parent_dir + '\\Data\\LMS\\dagdelen\\trp_trein1.dat', delimiter='\t')
trein2 = pd.read_csv(parent_dir + '\\Data\\LMS\\dagdelen\\trp_trein2.dat', delimiter='\t')
trein3 = pd.read_csv(parent_dir + '\\Data\\LMS\\dagdelen\\trp_trein3.dat', delimiter='\t')


carpas1 = pd.read_csv(parent_dir + '\\Data\\LMS\\dagdelen\\trp_pas1.dat', delimiter='\t')
carpas2 = pd.read_csv(parent_dir + '\\Data\\LMS\\dagdelen\\trp_pas2.dat', delimiter='\t')
carpas3 = pd.read_csv(parent_dir + '\\Data\\LMS\\dagdelen\\trp_pas3.dat', delimiter='\t')


ratio = car_final.tot / car_driver.tot
ratio3 = car_final3.tot / car_driver.tot
ratio2 = car_final2.tot / car_driver.tot


ratio = np.array(ratio)
ratio[ratio == np.inf] = np.nan

ratio3 = np.array(ratio3)
ratio3[ratio3 == np.inf] = np.nan

ratio2 = np.array(ratio2)
ratio2[ratio2 == np.inf] = np.nan

plt.hist(ratio, bins=np.arange(0, 1, 0.01), density=True)
plt.hist(ratio3, bins=np.arange(0, 1, 0.01), alpha=0.5, density=True);
plt.hist(ratio2, bins=np.arange(0, 1, 0.01), alpha=0.5, density=True);

# plt.hist(np.hstack([ratio, ratio2, ratio3]), bins=np.arange(0, 1, 0.01), alpha=0.5, density=True, cumulative=True);


car_final = car_final[(car_final.Orig.isin(np.arange(1, 1407))) & (car_final.Dest.isin(np.arange(1, 1407)))]
car_final2 = car_final2[(car_final2.Orig.isin(np.arange(1, 1407))) & (car_final2.Dest.isin(np.arange(1, 1407)))]
car_final3 = car_final3[(car_final3.Orig.isin(np.arange(1, 1407))) & (car_final3.Dest.isin(np.arange(1, 1407)))]


c1 = car_final.groupby('Orig').sum()
c2 = car_final2.groupby('Orig').sum()
c3 = car_final3.groupby('Orig').sum()


car_ratios = [c1, c2, c3]

day_parts = [[car_final, car_final2, car_final3],
             [carpas1, carpas2, carpas3],
             [trein1, trein2, trein3],
             [btm1, btm2, btm3]]


sum_day_parts = [[], [], [],[]]

for m in range(4):
    for d in range(3):

        day_parts[m][d] = day_parts[m][d][(day_parts[m][d].Orig.isin(np.arange(1, 1407))) & (day_parts[m][d].Dest.isin(np.arange(1, 1407)))]

        sum_day_parts[m].append(day_parts[m][d].groupby('Orig').sum().iloc[:, -1])


lms_orig

sum_day_parts[3][2]

## OViN day parts

ovin_modal = ovin[(~ovin.KHvm.isin([5, 8])) & (~ovin.KHvm.isnull())] # Delete trips with mopeds and 'other' or trips without mode

ovin_mor = ovin_modal[ovin_modal.KVertTijd.isin([3, 4])]
ovin_eve = ovin_modal[ovin_modal.KVertTijd.isin([9, 10])]
ovin_rest = ovin_modal[ovin_modal.KVertTijd.isin([1, 2, 5, 6, 7, 8, 11, 12, 13])]

ovin_list = [ovin_mor, ovin_rest, ovin_eve]

tb_data = zones[['ZONE_ID']]

Morning peak:  7 - 9
Evening peak: 16 - 18
Documentatie D1 5.3.2

!! Nog nakijken welke trips bij welk dagdeel indelen!!

# vert_tot = odin_modal[['VertZone', 'FactorV_final']].groupby('VertZone').sum()

m_list = [1, 2, 3, 4, 6, 7]

labels_vert = [['Car_driver_mor', 'Car_passenger_mor', 'Train_mor', 'BTM_mor', 'Bike_mor', 'Walk_mor'],
                ['Car_driver_rest', 'Car_passenger_rest', 'Train_rest', 'BTM_rest', 'Bike_rest', 'Walk_rest'],
                ['Car_driver_eve', 'Car_passenger_eve', 'Train_eve', 'BTM_eve', 'Bike_eve', 'Walk_eve']]

for i in range(3):

    vert_tot = ovin_list[i][['VertZone', 'FactorV_final']].groupby('VertZone').sum()

    for m in range(len(m_list)):

        split = ovin_list[i][ovin_list[i].KHvm == m_list[m]][['VertZone', 'FactorV_final']].groupby('VertZone').sum()
    	
        # split = np.round(split / vert_tot * 100, 2)

        tb_data = tb_data.merge(split, left_on='ZONE_ID', right_index=True, how='left')
        tb_data = tb_data.rename(columns={'FactorV_final': labels_vert[i][m]})



# ovin_list[i][ovin_list[i].KHvm == m_list[m]][['VertZone', 'FactorV_final']].groupby('VertZone').sum()

tb_data

## Create DU graph

ovin_tb = pd.read_csv(parent_dir + '\\Data\\New\\lms_zone_ovin_travel_behaviour_newF.csv', index_col=0)
lms_tb = pd.read_csv(parent_dir + '\\Data\\New\\lms_zone_lms_modal_split.csv', index_col=0)

lms_du = np.zeros((6, 6))
ovin_du = np.zeros((6, 6))
lms_part = np.zeros((6, 4, 3))
lms_ratio = np.zeros((6, 4, 3))
ovin_part = np.zeros((6, 4, 3))
ovin_ratio = np.zeros((6, 4, 3))

for du in range(1, 7):

    zone_list = zones[zones.deg_urba == du].ZONE_ID # Get all zones with certain DU
    
    # Determine percentagese for LMS
    lms_sum = lms_orig[lms_orig.ZONE_ID.isin(zone_list)].sum()[1:] # Get total trips of each mode
    lms_tot = lms_sum.sum()

    p = lms_sum / lms_tot * 100
    p.iloc[3] = p.iloc[3:5].sum() # Combine btm
    p = p.iloc[np.r_[0:4, 5:7]] # Remove metro row

    lms_du[du - 1, :] = p

    # Determine ratio dayparts
    for m in range(4):
        for c in range(3):
            if m == 3:

                lms_part[du - 1, m, c] = sum_day_parts[m][c][sum_day_parts[m][c].index.isin(zone_list)].sum() / lms_sum.iloc[m:m + 2].sum() * p.iloc[m]
                lms_ratio[du - 1, m, c] = sum_day_parts[m][c][sum_day_parts[m][c].index.isin(zone_list)].sum() / lms_sum.iloc[m:m + 2].sum()

            else:
                lms_part[du - 1, m, c] = sum_day_parts[m][c][sum_day_parts[m][c].index.isin(zone_list)].sum() / lms_sum.iloc[m] * p.iloc[m]
                lms_ratio[du - 1, m, c] = sum_day_parts[m][c][sum_day_parts[m][c].index.isin(zone_list)].sum() / lms_sum.iloc[m]



    # Determine percentages for OViN
    ovin_df = ovin_tb[ovin_tb.ZONE_ID.isin(zone_list)].iloc[:, np.r_[1:7, 29]] # Get right ovin data

    for i in range(6): # Loop over modes and calculate weighted average
        ovin_du[du - 1, i] = (ovin_df.iloc[:, i] * ovin_df.iloc[:, -1]).sum() / ovin_df.iloc[:, -1].sum()
    
        if i < 4:
            for c in range(3):
                ovin_df_part = np.array(tb_data[tb_data.ZONE_ID.isin(zone_list)].iloc[:, i + 1 + c * 6])
                ovin_df_part[ovin_df_part != ovin_df_part] = 0
                ovin_df_part = np.sum(ovin_df_part) / ovin_df.iloc[:, -1].sum()
                ovin_ratio[du - 1, i, c] = ovin_df_part / ovin_du[du - 1, i] * 100
                ovin_part[du - 1, i, c] = ovin_df_part 


tot_btm = lms_ratio[:, 3].sum(axis=1)
tot_ovin = ovin_ratio.sum(axis=2)

# Small rounding error, scale values to 1 for btm and OViN
for i in range(3):
    lms_ratio[:, 3, i] = lms_ratio[:, 3, i] / tot_btm

    ovin_ratio[:, :, i] = ovin_ratio[:, :, i] / tot_ovin

du_list = np.arange(1, 7)
mode_labels = ['Car driver', 'Car passenger', 'Train',
          'BTM', '(e)-Bike', 'Walking']

time_labels = ['Morning peak', 'Rest day', 'Evening peak']

f, ax = plt.subplots(1, 4)
f.set_figwidth(18)
f.set_figheight(6)

f.suptitle('Share of trips departing with a certain mode during part of the day',
           fontsize='xx-large')

color_idx = ['darkred', 'darkolivegreen', 'orangered', 'olivedrab', 'lightsalmon', 'yellowgreen']

for m in range(4):
    for c in range(3):

        if c > 0:
            b = lms_ratio[:, m, 0:c].sum(axis=1)
            bo = ovin_ratio[:, m, 0:c].sum(axis=1)
        else:
            b = None
            bo = None

        ax[m].bar(du_list - 0.2, ovin_ratio[:, m, c], label=f'{time_labels[c]} OViN', bottom=bo, width=0.35, color=color_idx[c * 2])
        ax[m].bar(du_list + 0.2, lms_ratio[:, m, c], label=f'{time_labels[c]} LMS', bottom=b, width=0.35,  color=color_idx[c * 2 + 1])


        ax[m].set_xticks(np.arange(1, 7))
        ax[m].set_xlabel('Degree of urbanisation')
        ax[m].set_title(mode_labels[m])
        ax[m].set_ylim(0, 1)

        ax[m].set_axisbelow(True)
        ax[m].grid(axis='y')

# ax[0].plot([], [], color='firebrick', label='OViN', linewidth=5)
# ax[0].plot([], [], color='forestgreen', label='LMS', linewidth=5)
# ax[0].plot([], [], color='steelblue', label='Difference OViN - LMS', linewidth=5)
handles, labels = ax[0].get_legend_handles_labels()
f.legend(handles, labels, loc='upper right', bbox_to_anchor=(0.9, 1.15))

f, ax = plt.subplots(2, 6)
f.set_figwidth(15)
f.set_figheight(8)


max_val = [np.max([lms_du.max(), ovin_du.max()]), 80]
min_val = [0, -50]
step = [5, 10]

ax[1, 5].plot([], [], color='firebrick', label='OViN', linewidth=5)
ax[1, 5].plot([], [], color='forestgreen', label='LMS', linewidth=5)
ax[1, 5].plot([], [], color='steelblue', label='Difference OViN - LMS', linewidth=5)
handles, labels = ax[1, 5].get_legend_handles_labels()


for i in range(2):

    for j in range(6):

        if i == 0:
            ax[0, j].bar(du_list - 0.2, ovin_du[:, j], color='firebrick', width=0.35)
            
            ax[0, j].bar(du_list + 0.2, lms_du[:, j], color='forestgreen', width=0.35)
            
            ax[1, j].bar(du_list, (-ovin_du[:, j] + lms_du[:, j]) / ovin_du[:, j] * 100, color='steelblue')

            ax[0, j].set_title(f'{mode_labels[j]}')
            ax[1, j].set_title(f'{mode_labels[j]}')
        
        if i == 1:
            ax[i, j].set_xlabel('Degree of urbanisation')

        if (j < 4) & (i == 0):
            for c in range(3):
                if c > 0:
                    b = lms_part[:, j, 0:c].sum(axis=1)
                    bo = ovin_part[:, j, 0:c].sum(axis=1) * 100
                else:
                    b = None
                    bo = None

                ax[0, j].bar(du_list - 0.2, ovin_part[:, j, c] * 100, label=f'{time_labels[c]} OViN', bottom=bo, width=0.35, color=color_idx[c * 2])
                ax[0, j].bar(du_list + 0.2, lms_part[:, j, c], label=f'{time_labels[c]} LMS', bottom=b, width=0.35,  color=color_idx[c * 2 + 1])

            

        

        ax[i, j].set_axisbelow(True)
        ax[i, j].grid(axis='y')
        ax[i, j].set_yticks(np.arange(min_val[i], max_val[i] + step[i], step[i]))
        ax[i, j].set_ylim(min_val[i], max_val[i] + step[i])
        ax[i, j].set_xticks(np.arange(1, 7))

ax[0, 1].legend(loc='upper left', bbox_to_anchor=(-1.8, 1.6))
ax[0, 0].set_ylabel('Percentage of mode users')
ax[1, 0].set_ylabel('Difference in mode use in %')



f.legend(handles, labels, loc='upper right', bbox_to_anchor=(0.9, 1.05))
f.suptitle('(Difference in) modal split LMS and OViN \nfor degree of urbanisation of the origin zone',
           fontsize='xx-large');
